+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # rux CHANGELOG
 
+
+
+## v0.7.1 - 2025-06-17
+
+* Fix file load time; clean up StreamingMessage [#47](https://github.com/rsanheim/rux-meta/pull/47)
+
+## v0.7.0 - 2025-06-17
+
+* Replace urfave/cli with Kong [#44](https://github.com/rsanheim/rux-meta/pull/44)
+* Convert all tests to use testify assertions [#45](https://github.com/rsanheim/rux-meta/pull/45)
+
+## v0.6.10 - 2025-06-15
+
+* Consolidating init [#43](https://github.com/rsanheim/rux-meta/pull/43)
+
+## v0.6.9 - 2025-06-12
+Lotsa internal cleanup
+* Extract execution logic and add Kong CLI experiment [#31](https://github.com/rsanheim/rux-meta/pull/31)
+* Add the integration test to CI (oops) [#41](https://github.com/rsanheim/rux-meta/pull/41)
+* get caching working for integration in CI [#40](https://github.com/rsanheim/rux-meta/pull/40)
+* Planning for auto config [#35](https://github.com/rsanheim/rux-meta/pull/35)
+* sort and clean up. [#39](https://github.com/rsanheim/rux-meta/pull/39)
+* Fix release thinger [#38](https://github.com/rsanheim/rux-meta/pull/38)
+
 ## v0.6.8 - 2025-06-12
 
 * Fix: update ci [#37](https://github.com/rsanheim/rux-meta/pull/37)